Refactor for pricing rule service

+++ src/com/coder73/PricingCalculator.java
@@ -1,5 +1,5 @@
 package com.coder73;
 
 public interface PricingCalculator {
-    int getTotal();
+    double getTotal();
 }
